safely repeating changes from add-cjs-loader-resolve to remember what I did and filter any WIP mess I may have forgotten about
