+++ TODO.md
@@ -4,3 +4,9 @@
 
 Copy any relevant changes from `add-cjs-loader-resolve`
 
+I forgot exactly where I was in `add-cjs-loader-resolve`
+Re-do the renaming and moving that I did in that branch.
+Then diff to see that I did it correctly.
+Avoid introducing any accidental behavioral changes.
+
+
